CareerMatch/backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sqlalchemy==1.4.51
psycopg2-binary==2.9.7
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pydantic==1.10.7
python-dotenv==1.0.0
